+++ foundation_utf8/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.2.0
+
+* Added `foundation.com.utf8.each_codepoint/2,3`
+
 # 1.1.0
 
 * Added `foundation.com.utf8.next_codepoint_length/2`